% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_rescaled_tpm.R
\name{average_rescaled_tpm}
\alias{average_rescaled_tpm}
\title{Average rescaled TPMs}
\usage{
average_rescaled_tpm(
  input,
  grp_df,
  sample_id = "sample_name",
  gene_id = "ensembl_gene_id"
)
}
\arguments{
\item{input}{numeric matrix of rescaled TPMs with gene IDs as rownames and sample IDs as colnames. Typically produced by rnaseqtools::rescale_tpm.}

\item{grp_df}{data.frame or tibble with column sample_id and one or more grouping columns.}

\item{sample_id}{character vector of length 1 representing sample ID. Must be a column name of grp_df. Sample IDs should be of the same type in colnames of input and sample_id column of group_df.}

\item{gene_id}{character vector of length 1 representing gene ID. Used to label row names of input.}
}
\value{
a matrix with rescaled TPMs averaged by groups.
}
\description{
Average rescaled TPMs
}
\examples{
# Define genes of interest
genes <- calcium_genes_ex \%>\%
  dplyr::filter(protein_complex == "SERCA") \%>\%
  dplyr::pull(ensembl_gene_id)

# Rescale TPMs for selected genes
resc_tpm <- rescale_tpm(input = bulk_tpm_ex, genes = genes)

# Average rescaled TPMs by disease status using a grp_df providing sample metadata
mean_resc_tpm <- average_rescaled_tpm(input = resc_tpm,
                                      grp_df = bulk_sample_metadata_ex,
                                      sample_id = donor_id,
                                      gene_id = ensembl_gene_id)

}
