% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_sum_rescaled_tpm.R
\name{average_sum_rescaled_tpm}
\alias{average_sum_rescaled_tpm}
\title{Average summed rescaled TPMs}
\usage{
average_sum_rescaled_tpm(
  input,
  grp_df,
  gs_df,
  sample_id = "sample_name",
  gene_id = "ensembl_gene_id"
)
}
\arguments{
\item{input}{numeric matrix of rescaled TPMs with gene IDs as rownames and sample IDs as colnames. Typically produced by rnaseqtools::rescale_tpm.}

\item{grp_df}{data.frame or tibble with column <sample_id> and one or more grouping columns.}

\item{gs_df}{data.frame or tibble with two columns: <gene_id> and a column representing gene sets.}

\item{sample_id}{character vector of length 1 or name representing sample ID. Must be a column name of grp_df. Sample IDs should be of the same type in colnames of input and column <sample_id> of grp_df.}

\item{gene_id}{character vector of length 1 or name representing gene ID. Gene IDs should be of the same type in rownames of input and column <gene_id> of gs_df.}
}
\value{
a matrix with rescaled TPMs summed by combination of sample ID and gene set and then averaged by groups.
}
\description{
Average summed rescaled TPMs
}
\examples{
# Define gs_df
gs_df <- calcium_genes_ex \%>\%
  dplyr::select(ensembl_gene_id, protein_complex) \%>\%
  dplyr::filter(!is.na(protein_complex))

# Pull genes of interest from gs_df
genes <- gs_df \%>\%
  dplyr::pull(ensembl_gene_id)

# Rescale TPMs for selected genes
resc_tpm <- rescale_tpm(input = bulk_tpm_ex, genes = genes)

# Sum rescaled TPMs by donor ID and protein complex using a gs_df providing gene metadata
# and then average summed rescaled TPMs by disease status using a grp_df providing sample metadata
mean_sum_resc_tpm <- average_sum_rescaled_tpm(input = resc_tpm,
                                              grp_df = bulk_sample_metadata_ex,
                                              gs_df = gs_df,
                                              sample_id = donor_id,
                                              gene_id = ensembl_gene_id)

}
