% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_ids2indices.R
\name{batch_ids2indices}
\alias{batch_ids2indices}
\title{Convert gene identifiers to indices for gene sets - multiple collections}
\usage{
batch_ids2indices(
  input,
  collections,
  identifiers,
  gene_id = "ensembl_gene_id_version"
)
}
\arguments{
\item{input}{data.frame or tibble with gene_id column and collections list-columns.}

\item{collections}{character vector of gene collection names. All elements must be column names of input.}

\item{identifiers}{character vector of gene identifiers.}

\item{gene_id}{character vector of length 1 representing gene ID. Must be a column name of input.}
}
\value{
a list of collection-specific sublists containing gene set-specific integer vectors of gene indices in the vector identifiers.
}
\description{
Convert gene identifiers to indices for gene sets - multiple collections
}
\examples{
### Do not run ###
# collections <- batch_ids2indices(input = msigdb_collection_table,
#                                  collections = collection_names,
#                                  identifiers = msigdb_collection_table$ensembl_gene_id,
#                                  gene_id = "ensembl_gene_id")

}
