% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{bulk_sample_metadata_ex}
\alias{bulk_sample_metadata_ex}
\title{Bulk RNA-seq sample metadata}
\format{
A tibble with 6 rows and 2 variables:
\describe{
\item{donor_id}{donor IDs matching the column names of the TPM matrix (donor1 to donor6)}
\item{treatment}{disease status (healthy or diseased)}
}
}
\usage{
bulk_sample_metadata_ex
}
\description{
Example sample metadata table matching the TPM matrix.
}
\keyword{datasets}
