% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{bulk_tpm_ex}
\alias{bulk_tpm_ex}
\title{Bulk-seq TPM matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10073 rows and 6 columns.
}
\usage{
bulk_tpm_ex
}
\description{
Example TPM matrix with column and row names representing donor IDs and ensembl gene IDs, respectively.
The matrix has 10073 rows and 6 columns.
}
\keyword{datasets}
