% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{calcium_genes_ex}
\alias{calcium_genes_ex}
\title{Calcium-related genes}
\format{
A tibble with 4 columns:
\describe{
\item{ensembl_gene_id}{ensembl gene ID (e.g.: ENSG00000156026)}
\item{gene_symbol}{gene symbol (e.g.: MCU)}
\item{protein_complex}{protein complex (e.g.: mitochondrial calcium uniporter MCU)}
\item{functional_pathway}{functional pathway (e.g.: mitochondrial calcium uptake))}
}
}
\usage{
calcium_genes_ex
}
\description{
Table listing genes involved in intracellular calcium handling.
}
\keyword{datasets}
