% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mean_tpm.R
\name{compute_mean_tpm}
\alias{compute_mean_tpm}
\title{Compute mean TPMs}
\usage{
compute_mean_tpm(
  input,
  gene_id = "ensembl_gene_id_version",
  sample_metadata,
  sample_id_var,
  group_id_var,
  ref_gene_ids = NULL,
  gene_groups = NULL,
  cnt_nm = "tpm"
)
}
\arguments{
\item{input}{data.frame or tibble with character columns <gene_id> and double columns representing sample-specific gene counts.}

\item{gene_id}{character vector of length 1 representing gene ID. Must be a column name of input.}

\item{sample_metadata}{data.frame or tibble with character / factor columns <sample_id_var> and <group_id_var>.}

\item{sample_id_var}{character vector of length 1 representing sample ID. Must be a column name of sample_metadata.}

\item{group_id_var}{character vector of length 1 representing biological group ID. Must be a column name of sample_metadata.}

\item{ref_gene_ids}{named character vector of gene IDs to use as reference. Values must be of the gene ID type specified in gene_id and names will be used in column names of output.}

\item{gene_groups}{list of named character vectors representing gene groups to use as reference.}

\item{cnt_nm}{ss}
}
\value{
ss
}
\description{
Compute mean TPMs
}
