% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{deseq2_results_ex}
\alias{deseq2_results_ex}
\title{DESeq2 results}
\format{
A tibble with 500 rows and 10 variables:
\describe{
\item{ensembl_gene_id_version}{ensembl gene ID with version number (e.g.: ENSG00000145335.17)}
\item{ensembl_gene_id}{ensembl gene ID (e.g.: ENSG00000145335)}
\item{gene_symbol}{gene symbol (e.g.: SNCA)}
\item{baseMean}{see ?DESeq2::results}
\item{log2FoldChange}{see ?DESeq2::results}
\item{lfcSE}{see ?DESeq2::results}
\item{stat}{see ?DESeq2::results}
\item{pvalue}{see ?DESeq2::results}
\item{padj}{see ?DESeq2::results}
\item{sign_log2fc_times_minus_log10pvalue}{gene ranking metric computed as sign(log2FoldChange) * -log(x = pvalue, base = 10)}
}
}
\usage{
deseq2_results_ex
}
\description{
Example DESeq2 results table produced by rnaseqtools::tidy_DESeqResults.
The table was augmented with columns for alternative gene IDs.
}
\keyword{datasets}
