% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{DESeqResults_ex}
\alias{DESeqResults_ex}
\title{DESeq2 DESeqResults}
\format{
An object of class \code{DESeqResults} with 500 rows and 6 columns.
}
\usage{
DESeqResults_ex
}
\description{
Example DESeqResults object produced by DESeq2::results.
}
\keyword{datasets}
