% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gencode_annotation.R
\name{download_gencode_annotation}
\alias{download_gencode_annotation}
\title{Download GENCODE annotation}
\usage{
download_gencode_annotation(
  species = "human",
  release = "latest",
  out_dir_path
)
}
\arguments{
\item{species}{character vector of length 1 representing species. Must be "human" or "mouse".}

\item{release}{"latest" or numeric vector of length 1 representing release number.}

\item{out_dir_path}{output directory.}
}
\value{
download requested GENCODE annotation to output directory.
}
\description{
Download GENCODE annotation
}
\examples{
# Download GENCODE annotation file
persistent_data_dir_path <- tools::R_user_dir(package = "rnaseqtools", which = "data")

download_gencode_annotation(species = "human",
                            release = 46,
                            out_dir_path = persistent_data_dir_path)

# Clean up
gencode_file_path <- paste0(persistent_data_dir_path,
                            "/",
                            "gencode.v46.primary_assembly.annotation.gtf.gz")

file.remove(gencode_file_path)

rm(persistent_data_dir_path, gencode_file_path)

}
