% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bioanalyzer_results.R
\name{extract_bioanalyzer_results}
\alias{extract_bioanalyzer_results}
\title{Extract Bioanalyzer results}
\usage{
extract_bioanalyzer_results(file)
}
\arguments{
\item{file}{a Results.csv file produced by Agilent 2100 Expert software.}
}
\value{
a tibble with summary of bioanalyzer results.
}
\description{
Extract Bioanalyzer results
}
\examples{
# Extract Bioanalyzer results
bio_res <- extract_bioanalyzer_results(file = rnaseqtools_example(file = "bio_res_ex.csv"))

}
