% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lmer_continuous.R
\name{fit_lmer_continuous}
\alias{fit_lmer_continuous}
\title{Linear mixed effects model for gene expression data}
\usage{
fit_lmer_continuous(
  sca,
  assay_name = "et",
  genes,
  gene_id,
  model_formula,
  coef_var,
  coef_test_lvl,
  cnt_scale = "log2",
  parallel = FALSE,
  ...
)
}
\arguments{
\item{sca}{SingleCellAssay object.}

\item{assay_name}{character vector of length 1 representing the assay to use; must be an assay name of \if{html}{\out{<sca>}}.}

\item{genes}{character vector of gene IDs to process; must be rownames of \if{html}{\out{<sca>}}.}

\item{gene_id}{character vector of length 1 representing gene ID. Used to name the gene ID column of output.}

\item{model_formula}{character vector of length 1 representing the RHS of the model formula. Must start with '~' and contain '|' for random effect.}

\item{coef_var}{character vector of length 1 representing the coefficient of interest. Must be a column name of colData(\if{html}{\out{<sca>}}).}

\item{coef_test_lvl}{character vector of length 1 representing the test level for the coefficient of interest. Must be a value of <coef_var>.}

\item{cnt_scale}{character vector of length 1 representing count scale. Either 'log2' or 'log'.}

\item{parallel}{logical vector of length 1 indicating if parallelisation should be enabled. If TRUE, use n detected cores - 1. Not implemented for windows.}

\item{...}{optional arguments passed to lmerTest::lmer.}
}
\value{
a tibble with lmerTest::lmer results.
}
\description{
Linear mixed effects model for gene expression data
}
\details{
For genes detected in every (or almost every) cell, MAST zlm(method = "glmer") does not compute fold change estimates because the discrete component of the two-part hurdle model cannot be fit.
fit_lmer_continuous() provides an alternative fold change estimation for those genes by fitting a linear mixed effects model for the continuous component only.
}
\examples{
# Construct SingleCellAssay object from sc_log2_tpm1p_ex and sc_sample_metadata_ex
sca <- MAST::FromMatrix(exprsArray = sc_log2_tpm1p_ex,
                        cData = as.data.frame(sc_sample_metadata_ex))

# Subset MAST glmer test results to only the genes with missing estimates
mast_zlm_glmer_failed <- mast_results_ex[is.na(mast_results_ex$log2FoldChange),]

# Fit mixed effect model for genes with missing estimates
results <- fit_lmer_continuous(sca = sca,
                               genes = mast_zlm_glmer_failed$ensembl_gene_id,
                               gene_id = "ensembl_gene_id",
                               model_formula = "~ treatment + (1 | donor_id)",
                               coef_var = "treatment",
                               coef_test_lvl = "treated",
                               parallel = FALSE)

}
