% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{gene_metadata_ex}
\alias{gene_metadata_ex}
\title{Gene metadata}
\format{
A tibble with 63140 rows and 5 variables:
\describe{
\item{ensembl_gene_id_version}{ensembl gene ID with version number (e.g.: ENSG00000145335.17)}
\item{ensembl_gene_id}{ensembl gene ID (e.g.: ENSG00000145335)}
\item{gene_symbol}{gene symbol (e.g.: SNCA)}
\item{gene_type}{gene type (e.g.: protein_coding)}
\item{chr_name}{chromosome name (e.g.: chr4))}
}
}
\usage{
gene_metadata_ex
}
\description{
Example gene metadata derived from GENCODE annotation - Human release 46.
It can be produced by successive calls to download_gencode_annotation, import_gencode_annotation and make_gene_metadata.
}
\keyword{datasets}
