% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_cols}
\alias{get_cols}
\title{Get columns from an RNA-seq dataset}
\usage{
get_cols(data, samples = NULL, cols = ".counts?$")
}
\arguments{
\item{data}{df rna-seq data to get columns from}

\item{samples}{df a samples df to subset the data to}

\item{cols}{Str to use as a search term}
}
\value{
data.frame
If there no matching columns in the supplied data frame
then NULL is returned
}
\description{
\code{get_cols} is a helper function to get either counts,
normalised counts or tpm from an RNA-seq data frame
}
\examples{
counts <- get_cols(data, samples, ".counts?$")
norm_counts <- get_cols(data, samples, ".normalised.counts?$")
tpm <- get_cols(data, samples, ".tpms?$")

}
