% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_counts}
\alias{get_counts}
\title{Get count data for an RNA-seq dataset}
\usage{
get_counts(data, samples = NULL, normalised = FALSE)
}
\arguments{
\item{data}{df rna-seq data to get counts from}

\item{samples}{df a samples df to subset the data to}

\item{normalised}{logical indicating whether to return
normalised counts or raw}
}
\value{
data.frame
}
\description{
\code{get_counts} returns a data.frame of counts,
either raw or normalised (default)
}
\examples{
norm_counts <- get_counts(data)
counts <- get_counts(data, normalised = FALSE)
counts_subset <- get_counts(data, samples)

}
