% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_gene_metadata}
\alias{get_gene_metadata}
\title{Get non count data for an RNA-seq dataset}
\usage{
get_gene_metadata(data)
}
\arguments{
\item{data}{df rna-seq data to get gene metadata from}
}
\value{
data.frame
}
\description{
\code{get_gene_metadata} returns a data.frame of metadata,
associated with the genes of an RNA-seq dataset
}
\examples{
non_count_data <- get_gene_metadata(data)

}
