% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msigdb_collections.R
\name{get_msigdb_collections}
\alias{get_msigdb_collections}
\title{Get MSigDB collections}
\usage{
get_msigdb_collections(species = "Homo sapiens", db_species = "Hs")
}
\arguments{
\item{species}{character vector of length 1 representing species for output genes. Must be a species_name returned by msigdbr::msigdbr_species.}

\item{db_species}{character vector of length 1 representing species abbreviation for database query. Must be "Hs" or "HS" for human and "Mm" or "MM" for mouse databases.}
}
\value{
a tibble with ensembl_gene_id column and collections list-columns. Each collection contains columns set_id and set_name.
}
\description{
Get MSigDB collections
}
\examples{
msigdb_collection_table = get_msigdb_collections()

}
