% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overlap.R
\name{get_overlap}
\alias{get_overlap}
\title{Get overlapping genes}
\usage{
get_overlap(input, set_name)
}
\arguments{
\item{input}{data.frame or tibble with set_name column and overlapGenes list-column, as produced by rnaseqtools::multi_fora.}

\item{set_name}{character vector of length 1 representing gene set name.}
}
\value{
a character vector of gene IDs representing the overlapping genes for the queried gene set.
}
\description{
Get overlapping genes
}
\examples{
# Get genes in overlap
overlap <- get_overlap(input = multi_fora_results_ex,
                       set_name = "HALLMARK_OXIDATIVE_PHOSPHORYLATION")

}
