% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_tpms}
\alias{get_tpms}
\title{Get TPM data for an RNA-seq dataset}
\usage{
get_tpms(data, samples = NULL)
}
\arguments{
\item{data}{df rna-seq data to get tpms from}

\item{samples}{df a samples df to subset the data to}
}
\value{
data.frame
If there is no tpm data in the supplied data frame
(i.e.) no column names containing "tpm" then NULL is returned
}
\description{
\code{get_tpm} returns a data.frame of tpms,
}
\examples{
tpms <- get_tpms(data)
tpms_subset <- get_tpms(data, samples)

}
