% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gencode_annotation.R
\name{import_gencode_annotation}
\alias{import_gencode_annotation}
\title{Import GENCODE annotation}
\usage{
import_gencode_annotation(file)
}
\arguments{
\item{file}{A GENCODE annotation file in GTF format.}
}
\value{
A tibble with GENCODE annotation.
}
\description{
Import GENCODE annotation
}
\examples{
# Download GENCODE annotation file
out_dir_path <- tools::R_user_dir(package = "rnaseqtools", which = "data")

download_gencode_annotation(species = "human",
                            release = 46,
                            out_dir_path = out_dir_path)

# Import GENCODE annotation from downloaded file
gencode_file_path <- paste0(out_dir_path,
                            "/",
                            "gencode.v46.primary_assembly.annotation.gtf.gz")

gencode_annotation <- import_gencode_annotation(file = gencode_file_path)

# Clean up
file.remove(gencode_file_path)

rm(out_dir_path, gencode_file_path, gencode_annotation)

}
