% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_salmon_meta_info.R
\name{import_salmon_meta_info}
\alias{import_salmon_meta_info}
\title{Import Salmon's metadata}
\usage{
import_salmon_meta_info(file)
}
\arguments{
\item{file}{A meta_info.json file produced by Salmon software.}
}
\value{
A tibble with key metrics extracted from input meta_info.json file.
}
\description{
Import Salmon's metadata
}
\examples{
# Import Salmon's run metadata
meta_info <- import_salmon_meta_info(file = rnaseqtools_example(file = "meta_info.json"))

}
