% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_rnaseq_data}
\alias{load_rnaseq_data}
\title{Load RNA-seq count data in from file}
\usage{
load_rnaseq_data(data_file, ...)
}
\arguments{
\item{data_file}{char, Name of the file to open}

\item{...}{Arguments passed to \code{\link[readr]{read_tsv}}}
}
\value{
tibble
}
\description{
\code{load_rnaseq_data} opens a file containing count data
and returns a tibble with column names adjusted to
take account of different column name formats.
}
\examples{
data <- load_rnaseq_data(data_file = 'all.tsv')

}
