% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_rnaseq_samples}
\alias{load_rnaseq_samples}
\title{Load RNA-seq samples file}
\usage{
load_rnaseq_samples(samples_file)
}
\arguments{
\item{data_file}{char, Name of the file to open}
}
\value{
tibble
}
\description{
\code{load_rnaseq_samples} opens a file containing sample
data and returns a tibble with column names adjusted to
take account of different column name formats.
}
\examples{
data <- load_rnaseq_samples(samples_file = 'samples.tsv')

}
