% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{log2_tpm1p_ex}
\alias{log2_tpm1p_ex}
\title{log2(TPM+1) matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 547 columns.
}
\usage{
log2_tpm1p_ex
}
\description{
Example log2(TPM+1) matrix with column and row names representing cell barcodes and ensembl gene IDs, respectively.
The matrix has 1000 rows and 547 columns.
}
\keyword{datasets}
