% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fgsea_pathways.R
\name{make_fgsea_pathways}
\alias{make_fgsea_pathways}
\title{Make fgsea pathways}
\usage{
make_fgsea_pathways(input, collections, gene_id = "ensembl_gene_id")
}
\arguments{
\item{input}{data.frame or tibble with gene_id column and collections list-columns. Each collection must contain column set_name.}

\item{collections}{character vector of gene collection names. All elements must be column names of input.}

\item{gene_id}{character vector of length 1 representing gene ID. Must be a column name of input.}
}
\value{
a list of collection-specific sublists containing gene set-specific character vectors of gene IDs.
}
\description{
Make fgsea pathways
}
\examples{
# Make MSigDB collection table
msigdb_collection_table = get_msigdb_collections()

# Define collections of interest
collection_names <- c("MSigDB_H",
                      "MSigDB_C2_CP:REACTOME",
                      "MSigDB_C2_CP:KEGG_LEGACY",
                      "MSigDB_C5_GO:BP")

# Make fgsea pathways
collections <- make_fgsea_pathways(input = msigdb_collection_table,
                                   collections = collection_names,
                                   gene_id = "ensembl_gene_id")

}
