% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_full_rank.R
\name{make_full_rank}
\alias{make_full_rank}
\title{Make matrix full rank}
\usage{
make_full_rank(model_matrix)
}
\arguments{
\item{model_matrix}{model matrix.}
}
\value{
the input matrix if already full rank.
A subset matrix if the input matrix is not full rank due to column(s) full of zeros.
An error if the input matrix is not full rank due to linearly dependent columns.
}
\description{
Make matrix full rank
}
\examples{

# Add genotype information to sample metadata
sample_metadata <- sc_sample_metadata_ex

sample_metadata$genotype <- ifelse(sample_metadata$donor_id == "donor1", "CTRL", "PD")

# Define model formula
model_formula <- "~ donor_id + treatment + genotype"

### Case when matrix is not full rank due to columns full of zeros

# Define model matrix (excluding donor2)
model_mtx <- model.matrix(object = formula(model_formula),
                          data = sample_metadata[sample_metadata$donor_id != "donor2",])

# Attempt to make model matrix full rank
message("Error captured:\n",
        attr(x = try(make_full_rank(model_matrix = model_mtx), silent = TRUE),
             which = "condition")$message)

### Case when matrix is not full rank due to linearly dependent columns

# Define model matrix (using full sample metadata)
model_mtx <- model.matrix(object = formula(model_formula),
                          data = sample_metadata)

# Attempt to make model matrix full rank
message("Error captured:\n",
        attr(x = try(make_full_rank(model_matrix = model_mtx), silent = TRUE),
             which = "condition")$message)

}
