% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gene_metadata.R
\name{make_gene_metadata}
\alias{make_gene_metadata}
\title{Make gene metadata}
\usage{
make_gene_metadata(input)
}
\arguments{
\item{input}{A GENCODE annotation tibble produced by import_gencode_annotation.}
}
\value{
A tibble with gene metadata.
}
\description{
Make gene metadata
}
\examples{
### Do not run ###
# gene_metadata <- make_gene_metadata(input = gencode_annotation)

}
