% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transcript_metadata.R
\name{make_transcript_metadata}
\alias{make_transcript_metadata}
\title{Make transcript metadata}
\usage{
make_transcript_metadata(input)
}
\arguments{
\item{input}{A GENCODE annotation data.frame or tibble as produced by import_gencode_annotation.}
}
\value{
A tibble with transcript metadata.
}
\description{
Make transcript metadata
}
\examples{
### Do not run ###
# transcript_metadata <- make_transcript_metadata(input = gencode_annotation)

}
