% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{mast_results}
\alias{mast_results}
\title{MAST results}
\format{
A tibble with 1000 rows and 10 variables:
\describe{
\item{ensembl_gene_id_version}{ensembl gene ID with version number (e.g.: ENSG00000145335.17)}
\item{ensembl_gene_id}{ensembl gene ID (e.g.: ENSG00000145335)}
\item{gene_symbol}{gene symbol (e.g.: SNCA)}
\item{Pr(>Chisq)}{see ?DESeq2::results}
\item{coef}{see ?DESeq2::results}
\item{ci.hi}{see ?DESeq2::results}
\item{ci.lo}{see ?DESeq2::results}
\item{padj}{adjusted p value computed as p.adjust(p = \verb{Pr(>Chisq)}, method = "BH")}
\item{log2fc}{log2 fold change computed as coef / log(x = 2, base = exp(1))}
\item{sign_log2fc_times_minus_log10pvalue}{gene ranking metric computed as sign(log2fc) * -log(x = \verb{Pr(>Chisq)}, base = 10)}
}
}
\usage{
mast_results
}
\description{
Example results table derived from MAST::zlm using the workflow described at \url{https://bioconductor.org/packages/release/bioc/vignettes/MAST/inst/doc/MAITAnalysis.html}.
The table was augmented with columns for alternative gene IDs, adjusted p value, log2 fold change and gene ranking metric.
}
\keyword{datasets}
