% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{mast_results_ex}
\alias{mast_results_ex}
\title{MAST results}
\format{
A tibble with 1000 rows and 11 variables:
\describe{
\item{ensembl_gene_id_version}{ensembl gene ID with version number (e.g.: ENSG00000145335.17)}
\item{ensembl_gene_id}{ensembl gene ID (e.g.: ENSG00000145335)}
\item{gene_symbol}{gene symbol (e.g.: SNCA)}
\item{Pr(>Chisq)}{see ?MAST::\verb{summary,ZlmFit-method}}
\item{coef}{see ?MAST::\verb{summary,ZlmFit-method}}
\item{ci.hi}{see ?MAST::\verb{summary,ZlmFit-method}}
\item{ci.lo}{see ?MAST::\verb{summary,ZlmFit-method}}
\item{log2FoldChange}{log2 fold change computed as coef / log(x = 2, base = exp(1))}
\item{lfcSE}{log2 fold change standard error computed from upper and lower bounds of confidence interval and level of confidence coefficient}
\item{padj}{adjusted p value computed as p.adjust(p = \verb{Pr(>Chisq)}, method = "BH")}
\item{sign_log2fc_times_minus_log10pvalue}{gene ranking metric computed as sign(log2FoldChange) * -log(x = \verb{Pr(>Chisq)}, base = 10)}
}
}
\usage{
mast_results_ex
}
\description{
Example MAST results table produced by rnaseqtools::tidy_summaryZlmFit.
The table was augmented with columns for alternative gene IDs.
}
\keyword{datasets}
