% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{msa_ex}
\alias{msa_ex}
\title{msa MsaAAMultipleAlignment}
\format{
An object of class \code{MsaAAMultipleAlignment} with 8 rows and 171 columns.
}
\usage{
msa_ex
}
\description{
Example MsaAAMultipleAlignment object produced by msa::msa.
}
\keyword{datasets}
