% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_to_tibble.R
\name{msa_to_tibble}
\alias{msa_to_tibble}
\title{MsaAAMultipleAlignment to tibble}
\usage{
msa_to_tibble(input)
}
\arguments{
\item{input}{MsaAAMultipleAlignment object produced by msa::msa.}
}
\value{
a tibble with columns name and sequence.
}
\description{
MsaAAMultipleAlignment to tibble
}
\examples{
# Extract multiple sequence alignment from MsaAAMultipleAlignment object
msa_df <- msa_to_tibble(input = msa_ex)

}
