% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{multi_fgsea_results}
\alias{multi_fgsea_results}
\title{Multi fgsea results}
\format{
A tibble with 32 rows and 12 variables:
\describe{
\item{collection}{collection name (e.g.: MSigDB_C2_CP:REACTOME)}
\item{set_name}{gene set name (e.g.: REACTOME_PHOSPHOLIPID_METABOLISM)}
\item{pval}{see ?fgsea::fgsea}
\item{padj}{see ?fgsea::fgsea}
\item{log2err}{see ?fgsea::fgsea}
\item{ES}{see ?fgsea::fgsea}
\item{NES}{see ?fgsea::fgsea}
\item{size}{see ?fgsea::fgsea}
\item{leadingEdge}{see ?fgsea::fgsea}
\item{n_leadingEdge}{number of genes in leading edge)}
\item{n_leadingEdge_over_total}{number of genes in leading edge over number of genes in gene set}
\item{pct_leadingEdge}{percentage of gene set being leading edge}
}
}
\usage{
multi_fgsea_results
}
\description{
Example results table produced by rnaseqtools::multi_fgsea.
}
\keyword{datasets}
