% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{multi_fora_results}
\alias{multi_fora_results}
\title{Multi fora results}
\format{
A tibble with 49 rows and 9 variables:
\describe{
\item{collection}{collection name (e.g.: MSigDB_H)}
\item{set_name}{gene set name (e.g.: HALLMARK_OXIDATIVE_PHOSPHORYLATION)}
\item{pval}{see ?fgsea::fora}
\item{padj}{see ?fgsea::fora}
\item{overlap}{see ?fgsea::fora}
\item{size}{see ?fgsea::fora}
\item{overlapGenes}{vector with overlapping genes}
\item{n_overlapGenes}{number of overlapping genes over number of genes in gene set}
\item{pct_overlapGenes}{percentage of gene set being overlapping genes}
}
}
\usage{
multi_fora_results
}
\description{
Example results table produced by rnaseqtools::multi_fora.
}
\keyword{datasets}
