% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ids2indices.R
\name{multi_ids2indices}
\alias{multi_ids2indices}
\title{Convert gene identifiers to indices for gene sets - multiple collections}
\usage{
multi_ids2indices(
  input,
  collections,
  identifiers,
  gene_id = "ensembl_gene_id_version"
)
}
\arguments{
\item{input}{data.frame or tibble with gene_id column and collections list-columns.}

\item{collections}{character vector of gene collection names. All elements must be column names of input.}

\item{identifiers}{character vector of gene identifiers.}

\item{gene_id}{character vector of length 1 representing gene ID. Must be a column name of input.}
}
\value{
a list of collection-specific sublists containing gene set-specific integer vectors of gene indices in the vector identifiers.
}
\description{
Convert gene identifiers to indices for gene sets - multiple collections
}
\examples{
# Make MSigDB collection table
msigdb_collection_table = get_msigdb_collections()

# Define collections of interest
collection_names <- c("MSigDB_H", "MSigDB_C2_CP:REACTOME", "MSigDB_C2_CP:KEGG", "MSigDB_C5_GO:BP")

# Extract gene identifiers from log2 COUNT+1 matrix
identifiers <- rownames(log2_tpm1p)

# Get lists of indices for selected collections
collections <- multi_ids2indices(input = msigdb_collection_table,
                                 collections = collection_names,
                                 identifiers = identifiers,
                                 gene_id = "ensembl_gene_id")

}
