% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_limma_gsea.R
\name{multi_limma_gsea}
\alias{multi_limma_gsea}
\title{Run limma gene set tests on multiple collections}
\usage{
multi_limma_gsea(
  y,
  input,
  collections,
  design,
  contrast,
  gene_id = "ensembl_gene_id",
  min_set_size = 1,
  max_set_size = dim(y)[[1]] - 1,
  padj_threshold = Inf,
  limma_test = "fry",
  ...
)
}
\arguments{
\item{y}{matrix giving log-expression or log-ratio values with rownames corresponding to gene_id.}

\item{input}{data.frame or tibble with gene_id column and collections list-columns.}

\item{collections}{character vector of gene collection names. All elements must be column names of input.}

\item{design}{design matrix.}

\item{contrast}{contrast for which the test is required. Can be an integer specifying a column of design, or the name of a column of design, or a numeric contrast vector of length equal to the number of columns of design.}

\item{gene_id}{character vector of length 1 representing gene ID. Must be a column name of input.}

\item{min_set_size}{minimal size of a gene set to test. All gene sets below the threshold are excluded.}

\item{max_set_size}{maximal size of a gene set to test. All gene sets above the threshold are excluded.}

\item{padj_threshold}{padj threshold. All gene sets with padj equal or above the threshold are filtered out.}

\item{limma_test}{character vector of length 1 representing limma gene set test.}

\item{...}{optional arguments passed to limma test.}
}
\value{
a tibble with results of limma test runs on multiple collections.
}
\description{
Run limma gene set tests on multiple collections
}
\examples{
# Make MSigDB collection table
msigdb_collection_table = get_msigdb_collections()

# Define collections of interest
collection_names <- c("MSigDB_H", "MSigDB_C2_CP:REACTOME", "MSigDB_C2_CP:KEGG", "MSigDB_C5_GO:BP")

# Extract gene identifiers from log2 COUNT+1 matrix
identifiers <- rownames(log2_tpm1p)

# Get lists of indices for selected collections
collections <- multi_ids2indices(input = msigdb_collection_table,
                                 collections = collection_names,
                                 identifiers = identifiers,
                                 gene_id = "ensembl_gene_id")

}
