% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{normalise_counts}
\alias{normalise_counts}
\title{Adjust RNAseq count data for library size}
\usage{
normalise_counts(rnaseq_data, samples)
}
\arguments{
\item{rnaseq_data}{data frame of rnaseq data. Must contain columns labelled
"{sample} count".}

\item{samples}{data frame of sample metadata. Must included a `condition` column.}
}
\value{
tibble containing the original data and the normalised count columns added
}
\description{
`normalise_counts()` takes RNAseq count data and produces normalised counts
that are adjusted for library size (sequencing depth). The normalised count
columns are added to the supplied data frame with column names that are
the original column names with " normalised count" appended.
}
\examples{

rnaseq_data <- test_all_data[ , !grepl("normalised", colnames(test_all_data)) ]
rnaseq_data <- normalise_counts(rnaseq_data, samples_data)

}
