% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_averaged_rescaled_tpm.R
\name{plot_averaged_rescaled_tpm}
\alias{plot_averaged_rescaled_tpm}
\title{Plot averaged rescaled TPMs}
\usage{
plot_averaged_rescaled_tpm(
  input,
  gene_id = "gene_name",
  grp_id = "genotype",
  error_bar = "sd",
  y_label = "\% of gene set",
  title = NULL,
  colors = c("#5d76cb", "#ff9e00", "#ca3767", "#38a3a5", "#f49cbb", "#858ae3", "#ffd500",
    "#2ba84a", "#fc6dab", "#7637da")
)
}
\arguments{
\item{input}{data.frame or tibble with character / factor columns <gene_id> and <grp_id> and numeric columns rescaled_tpm, mean_rescaled_tpm and <error_bar>_rescaled_tpm. Typically produced by rnaseqtools::average_rescaled_tpm.}

\item{gene_id}{character vector of length 1 or name representing gene ID. Must be a column name of input.}

\item{grp_id}{character vector of length 1 or name representing sample groups to display on x axis. Must be a column name of input.}

\item{error_bar}{character vector of length 1 or name representing metric to use for drawing error bars. Must be one of sd or sem such that <error_bar>_rescaled_tpm is a column name of input.}

\item{y_label}{character vector of length 1 representing x axis label.}

\item{title}{character vector of length 1 representing plot title.}

\item{colors}{character vector of colors to use for column fill based on <gene_id>.}
}
\value{
a ggplot object with averaged rescaled TPMs split by sample group. Labels express transcript abundance as a percentage of all transcripts for each group.
}
\description{
Plot averaged rescaled TPMs
}
\examples{
# Define genes of interest
genes <- calcium_genes_ex \%>\%
  dplyr::filter(protein_complex == "SERCA") \%>\%
  dplyr::pull(ensembl_gene_id)

# Rescale TPMs for selected genes
resc_tpm <- rescale_tpm(input = bulk_tpm_ex, genes = genes)

# Average rescaled TPMs by disease status using a grp_df providing sample metadata
mean_resc_tpm <- average_rescaled_tpm(input = resc_tpm,
                                      grp_df = bulk_sample_metadata_ex,
                                      sample_id = donor_id,
                                      gene_id = ensembl_gene_id)

# Add gene symbols to averaged rescaled TPMs
mean_resc_tpm <- mean_resc_tpm \%>\%
  dplyr::mutate(disease_status = forcats::fct_relevel(disease_status, c("healthy", "diseased"))) \%>\%
  dplyr::left_join(gene_metadata_ex[, c("ensembl_gene_id", "gene_symbol")],
                   by = dplyr::join_by(ensembl_gene_id))

# Plot averaged rescaled TPMs by disease status
plot_averaged_rescaled_tpm(input = mean_resc_tpm,
                           gene_id = gene_symbol,
                           grp_id = disease_status,
                           error_bar = sd,
                           y_label = "\% of gene set",
                           title = "SERCA protein complex")

}
