% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes_prop_in_leading_edges.R
\name{plot_genes_prop_in_leading_edges}
\alias{plot_genes_prop_in_leading_edges}
\title{Plot genes proportions in leading edges}
\usage{
plot_genes_prop_in_leading_edges(
  input,
  genes,
  x_lab = "gene set",
  gene_group_name = "searched",
  collection_levels = NULL,
  collection_colors = NULL
)
}
\arguments{
\item{input}{data.frame or tibble with columns collection and n_leadingEdge and list-column leadingEdge, as produced by rnaseqtools::multi_fgsea.}

\item{genes}{character vector of gene IDs to search in leading edges. Must be the same type of gene IDs as in the leadingEdge list-column.}

\item{x_lab}{character vector of length 1 representing the label of x axis.}

\item{gene_group_name}{character vector of length 1 representing the name of the genes group. Will be used in y axis label.}

\item{collection_levels}{character vector with collection levels in desired order.}

\item{collection_colors}{character vector with collection level colors in desired order.}
}
\value{
a ggplot object with genes proportions in leading edges for the selected genes.
}
\description{
Plot genes proportions in leading edges
}
\examples{
# Make MSigDB collection table
msigdb_collection_table = get_msigdb_collections()

# Define collection of interest
collection_name <- "MSigDB_C2_CP:KEGG_LEGACY"

# Define gene set name of interest
set_name <- "KEGG_RIBOSOME"

# Get genes in gene set of interest
genes <- get_genes_in_collection(input = msigdb_collection_table,
                                 collection = collection_name,
                                 set_name = set_name,
                                 gene_ids = "ensembl_gene_id")

genes <- genes$ensembl_gene_id

# Plot genes proportions in leading edges for the selected genes
plot_genes_prop_in_leading_edges(input = multi_fgsea_results_ex,
                                 genes = genes,
                                 x_lab = "gene set",
                                 gene_group_name = set_name,
                                 collection_levels = c("MSigDB_C2_CP:REACTOME",
                                                       "MSigDB_C2_CP:KEGG_LEGACY",
                                                       "MSigDB_C5_GO:BP"),
                                 collection_colors = c("#5d76cb",
                                                       "#29a655",
                                                       "#ca3767"))

}
