% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lfc.R
\name{plot_lfc}
\alias{plot_lfc}
\title{Plot log2 fold changes}
\usage{
plot_lfc(
  input,
  genes,
  gene_id = "ensembl_gene_id",
  pretty_gene_id = NULL,
  ref_cond = "ref cond.",
  test_cond = "test cond.",
  key_genes = NULL,
  key_genes_name = "leading edge",
  title = NULL,
  desc_lfc = TRUE,
  xintercepts = NULL,
  padj_lab = TRUE,
  lab_nudge = 0.02
)
}
\arguments{
\item{input}{data.frame or tibble with character columns <gene_id> and <pretty_gene_id>, and double columns log2FoldChange, lfcSE and padj.}

\item{genes}{character vector of gene IDs for which log2 fold changes should be displayed. Must be of the gene ID type specified in gene_id.}

\item{gene_id}{character vector of length 1 or name representing gene ID. Must be a column name of input.}

\item{pretty_gene_id}{character vector of length 1 or name representing alternative gene ID to use for display. Must be a column name of input.}

\item{ref_cond}{character vector of length 1 or name representing the reference biological condition.}

\item{test_cond}{character vector of length 1 or name representing the test biological condition.}

\item{key_genes}{character vector of gene IDs representing key genes among genes. Must be of the gene ID type specified in gene_id.}

\item{key_genes_name}{character vector of length 1 representing the name of the key genes group.}

\item{title}{character vector of length 1 representing plot title.}

\item{desc_lfc}{logical vector of length 1 indicating if log2 fold changes should be ordered in descending order.}

\item{xintercepts}{numeric vector of x positions where vertical lines should be drawn. Passed to ggplot2::geom_vline.}

\item{padj_lab}{logical vector of length 1 indicating if padj values should be displayed.}

\item{lab_nudge}{numerical vector of length 1 representing the amount of horizontal nudge to use for padj labels.}
}
\value{
a ggplot object with log2 fold changes for the selected genes in test vs reference conditions.
}
\description{
Plot log2 fold changes
}
\examples{
# Extract the gene IDs of top 25 upregulated genes with smallest padj
genes <- deseq2_results_ex \%>\%
  dplyr::filter(log2FoldChange > 0) \%>\%
  dplyr::arrange(padj) \%>\%
  dplyr::pull(ensembl_gene_id) \%>\%
  head(25)

# Plot their log2 fold changes
plot_lfc(input = deseq2_results_ex,
         genes = genes,
         gene_id = "ensembl_gene_id",
         pretty_gene_id = "gene_symbol",
         ref_cond = "untreated",
         test_cond = "treated",
         desc_lfc = FALSE,
         xintercepts = seq(0.5, 2, 0.5),
         padj_lab = TRUE,
         lab_nudge = 0.02)

}
