% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean_tpm.R
\name{plot_mean_tpm}
\alias{plot_mean_tpm}
\title{Plot mean TPMs}
\usage{
plot_mean_tpm(
  input,
  genes,
  gene_id = "ensembl_gene_id",
  pretty_gene_id = NULL,
  facet_var = NULL,
  color_var = NULL,
  expr_quartile = FALSE,
  ref_gene = NULL,
  fill_colors = c("#0077b6", "#a4133c", "#d3d3d3"),
  outline_colors = c("#5d76cb", "#ff9e00", "#29a655", "#ca3767", "#38a3a5"),
  title = NULL,
  x_label = "mean TPM",
  cnt_nm = "tpm"
)
}
\arguments{
\item{input}{data.frame or tibble with character columns <gene_id> and <pretty_gene_id> and double columns mean_<cnt_nm> and sem_<cnt_nm>.}

\item{genes}{character vector of gene IDs for which mean TPM should be be displayed. Must be of the gene ID type specified in gene_id.}

\item{gene_id}{character vector of length 1 or name representing gene ID. Must be a column name of input.}

\item{pretty_gene_id}{character vector of length 1 or name representing alternative gene ID to use for display. Must be a column name of input.}

\item{facet_var}{character vector of length 1 or name representing categorical variable to facet_wrap and fill-code by. Must be a column name of input.}

\item{color_var}{character vector of length 1 or name representing categorical variable to outline color-code by. Must be a column name of input.}

\item{expr_quartile}{logical vector of length 1 indicating if expression quartile labels should be displayed.}

\item{ref_gene}{character vector of length 1 representing the reference gene to use in labels. Ignored if expr_quartile is FALSE. mean_<cnt_nm>\emph{pct}<ref_gene> must be a column name of input.}

\item{fill_colors}{character vector of colors to use for column fill based on facet_var. If facet_var is NULL, the first element of fill_colors is used.}

\item{outline_colors}{character vector of colors to use for column outline based on color_var. If color_var is NULL, outline_colors is ignored and black outlines are drawn.}

\item{title}{character vector of length 1 representing plot title.}

\item{x_label}{character vector of length 1 representing x axis label.}

\item{cnt_nm}{character vector of length 1 representing the count name used in input.}
}
\value{
a ggplot object with mean TPMs for the selected genes and optionally expression quartile labels.
}
\description{
Plot mean TPMs
}
\examples{
# Convert TPM matrix to tibble with gene IDs as column
tpm_df <- tibble::as_tibble(x = bulk_tpm_ex, rownames = "ensembl_gene_id")

# Define reference genes
ref_gene_ids <- gene_metadata_ex \%>\%
  dplyr::filter(gene_symbol \%in\% c("ACTB", "GAPDH")) \%>\%
  dplyr::pull(ensembl_gene_id, gene_symbol)

# Compute mean TPMs per disease status
mean_tpm <- compute_mean_tpm(input = tpm_df,
                             gene_id = "ensembl_gene_id",
                             sample_metadata = bulk_sample_metadata_ex,
                             sample_id_var = "donor_id",
                             group_id_var = "disease_status",
                             ref_gene_ids = ref_gene_ids,
                             gene_groups = NULL,
                             cnt_nm = "tpm")

# Add gene symbols to mean TPM table
mean_tpm <- mean_tpm \%>\%
  dplyr::mutate(disease_status = forcats::fct_relevel(disease_status, c("healthy", "diseased"))) \%>\%
  dplyr::left_join(gene_metadata_ex[, c("ensembl_gene_id", "gene_symbol")],
                   by = dplyr::join_by(ensembl_gene_id))

# Define genes to display
genes <- gene_metadata_ex[match(x = c("SNCA", "TH", "SNAP25", "MAP2", "AQP4"),
                                table = gene_metadata_ex$gene_symbol),
                          "ensembl_gene_id"][[1]]

# Plot mean TPM for genes of interest
plot_mean_tpm(input = mean_tpm,
              genes = genes,
              gene_id = ensembl_gene_id,
              pretty_gene_id = gene_symbol,
              facet_var = disease_status,
              expr_quartile = TRUE,
              ref_gene = "ACTB",
              fill_colors = c("#0077b6", "#a4133c"))

}
