% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_medians.R
\name{plot_medians}
\alias{plot_medians}
\title{Plot the median expression value for a set of genes using a counts object and annotation dataframe}
\usage{
\code{plot_medians(
  counts, module,
  annotation, id,
  group, timepoint = NULL,
  palette, sig_comparison = NULL,
  ...)}
}
\arguments{
\item{counts}{Matrix or dataframe of counts or batch corrected residuals.  The dimension corresponding to genes must have dimnames, which will be matched to the gene identifiers in \code{gene_sets}. The dimension corresponding to library or sample IDs must have dimnames, which will be matched to the sample identifiers in \code{annotation}.}

\item{module}{The module or gene set, as a character vector}

\item{annotation}{Dataframe where libraries that should be included in plot are in rows and phenotypic/quality characteristics are in columns.}

\item{id}{Name of the column in \code{annotation} object corresponding to the library IDs used to name samples in the \code{counts} object.}

\item{group}{Name of the column in \code{annotation} object corresponding to experimental groups that should be plotted.}

\item{timepoint}{If NULL, data is not expected to be shown over time (default). Otherwise, the name of the column in \code{annotation} that will contain the existing time information.}

\item{sig_comparison}{Argument to pass on to geom_signif.  This should be a list of length-2 vectors that correspond to the group names in a pairwise comparison.  These names must be present in the \code{group} column of the \code{annotation} object.}

\item{palette}{The discrete color palette to be used for groups.  Defaults to a 6-color scheme.}

\item{...}{Optional parameters passed to \code{gene_set_median}.}
}
\value{
A plot showing the median expression values separated by experimental group, with either one element for each library in \code{counts} (if time series data was not collected) or one element for each group at each time point.
}
\description{
This function allows you to express your love of cats.
}
