% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_msa.R
\name{plot_msa}
\alias{plot_msa}
\title{Plot MSA}
\usage{
plot_msa(input)
}
\arguments{
\item{input}{data.frame or tibble with character columns name and sequence, as produced by rnaseqtools::msa_to_tibble.}
}
\value{
a ggplot object with multiple sequence alignment.
}
\description{
Plot MSA
}
\examples{
# Extract multiple sequence alignment from MsaAAMultipleAlignment object
msa_df <- msa_to_tibble(input = msa_ex)

# Plot multiple sequence alignment
plot_msa(input = msa_df)

}
