% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pct_exp.R
\name{plot_pct_exp}
\alias{plot_pct_exp}
\title{Plot percentage expression}
\usage{
plot_pct_exp(
  input,
  genes,
  gene_id = "ensembl_gene_id",
  pretty_gene_id = "gene_symbol",
  ref_cond = "ref cond.",
  test_cond = "test cond.",
  title = NULL,
  fill_colors = c("#ca3767", "#5d76cb")
)
}
\arguments{
\item{input}{data.frame or tibble with character columns <gene_id> and <pretty_gene_id>, and double column log2FoldChange.}

\item{genes}{character vector of gene IDs for which percentage expression in test vs reference conditions should be displayed. Must be of the gene ID type specified in gene_id.}

\item{gene_id}{character vector of length 1 representing gene ID. Must be a column name of input.}

\item{pretty_gene_id}{character vector of length 1 representing alternative gene ID to use for display. Must be a column name of input.}

\item{ref_cond}{character vector of length 1 representing the reference biological condition.}

\item{test_cond}{character vector of length 1 representing the test biological condition.}

\item{title}{character vector of length 1 representing plot title.}

\item{fill_colors}{character vector of length 2 representing the colors to use for reference and test conditions, respectively.}
}
\value{
a ggplot object with percentage expression in test relative to reference conditions for the selected genes.
}
\description{
Plot percentage expression
}
\examples{
# Extract the gene IDs of top 25 downregulated genes with smallest padj
genes <- deseq2_results_ex \%>\%
  dplyr::filter(log2FoldChange < 0) \%>\%
  dplyr::arrange(padj) \%>\%
  dplyr::pull(ensembl_gene_id) \%>\%
  head(25)

# Plot their percentage expression in test relative to reference conditions
plot_pct_exp(input = deseq2_results_ex,
             genes = genes,
             gene_id = "ensembl_gene_id",
             pretty_gene_id = "gene_symbol",
             ref_cond = "untreated",
             test_cond = "treated",
             title = NULL,
             fill_colors = c("#5d76cb", "#ca3767"))

}
