% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_tpm.R
\name{rescale_tpm}
\alias{rescale_tpm}
\title{Rescale TPMs}
\usage{
rescale_tpm(input, genes)
}
\arguments{
\item{input}{numeric matrix of TPMs with gene IDs as rownames and sample IDs as colnames.}

\item{genes}{character vector representing gene IDs to subset for TPM rescaling. Must be of the same gene ID type as rownames(input).}
}
\value{
a matrix with rescaled TPMs for the selected genes.
}
\description{
Rescale TPMs
}
\examples{
# Define genes of interest
genes <- calcium_genes_ex \%>\%
  dplyr::filter(protein_complex == "SERCA") \%>\%
  dplyr::pull(ensembl_gene_id)

# Rescale TPMs for selected genes
resc_tpm <- rescale_tpm(input = bulk_tpm_ex, genes = genes)

}
