% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_example.R
\name{rnaseqtools_example}
\alias{rnaseqtools_example}
\title{Get path to rnaseqtools example}
\usage{
rnaseqtools_example(file = NULL)
}
\arguments{
\item{file}{name of file. If NULL, the example files will be listed.}
}
\description{
rnaseqtools comes bundled with a number of sample files in its inst/extdata directory. This function make them easy to access.
}
\examples{
rnaseqtools_example()

rnaseqtools_example(file = "meta_info_ex.json")

}
