% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{run_pca}
\alias{run_pca}
\title{Run PCA}
\usage{
run_pca(
  input,
  pcs = 1:2,
  ntop = NULL,
  center = T,
  scale. = T,
  rownames = "barcode"
)
}
\arguments{
\item{input}{numeric matrix or data.frame which provides the data for the principal components analysis.}

\item{pcs}{principal components to return.}

\item{ntop}{number of top variable features to use for PCA.}

\item{center}{logical value indicating whether the variables should be shifted to be zero centered.}

\item{scale.}{logical value indicating whether the variables should be scaled to have unit variance before the analysis takes place.}

\item{rownames}{character vector of length 1 representing row names.}
}
\value{
a tibble with PCA results.
}
\description{
Run PCA
}
\examples{
# Run PCA
pca_results <- run_pca(input = sc_log2_tpm1p_ex,
                       pcs = 1:10)

}
