% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{sc_cnt_ex}
\alias{sc_cnt_ex}
\title{scRNA-seq raw count matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 500 rows and 547 columns.
}
\usage{
sc_cnt_ex
}
\description{
Example raw count matrix with column and row names representing cell barcodes and ensembl gene IDs, respectively.
The matrix has 500 rows and 547 columns.
}
\keyword{datasets}
