% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{sc_sample_metadata_ex}
\alias{sc_sample_metadata_ex}
\title{scRNA-seq sample metadata}
\format{
A tibble with 547 rows and 3 variables:
\describe{
\item{barcode}{cell barcodes matching the column names of the log2(TPM+1) matrix}
\item{donor_id}{donor ID (donor1, donor2 or donor3)}
\item{treatment}{treatment status (untreated or treated)}
}
}
\usage{
sc_sample_metadata_ex
}
\description{
Example sample metadata table matching the log2(TPM+1) matrix.
}
\keyword{datasets}
