% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_genes_in_leading_edges.R
\name{search_genes_in_leading_edges}
\alias{search_genes_in_leading_edges}
\title{Search genes in leading edges}
\usage{
search_genes_in_leading_edges(input, genes, group_name = "searched")
}
\arguments{
\item{input}{data.frame or tibble with columns collection and n_leadingEdge and list-column leadingEdge, as produced by rnaseqtools::multi_fgsea.}

\item{genes}{character vector of gene IDs to search in leading edges. Must be the same type of gene IDs as in the leadingEdge list-column.}

\item{group_name}{character vector of length 1 representing group name for the searched genes.}
}
\value{
the input table augmented with gene search results.
}
\description{
Search genes in leading edges
}
\examples{
# Get leading edge for KEGG_RIBOSOME gene set
genes <- get_leading_edge(input = multi_fgsea_results_ex,
                          set_name = "KEGG_RIBOSOME")

# Search genes representing KEGG_RIBOSOME leading edge in all leading edges
search_results <- search_genes_in_leading_edges(input = multi_fgsea_results_ex,
                                                genes = genes,
                                                group_name = "ribosome")

}
