% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_genes_in_overlaps.R
\name{search_genes_in_overlaps}
\alias{search_genes_in_overlaps}
\title{Search genes in overlaps}
\usage{
search_genes_in_overlaps(input, genes, group_name = "searched")
}
\arguments{
\item{input}{data.frame or tibble with columns collection and overlap and list-column overlapGenes, as produced by rnaseqtools::multi_fora.}

\item{genes}{character vector of gene IDs to search in overlaps. Must be the same type of gene IDs as in the overlapGenes list-column.}

\item{group_name}{character vector of length 1 representing group name for the searched genes.}
}
\value{
the input table augmented with gene search results.
}
\description{
Search genes in overlaps
}
\examples{
# Get overlap for KEGG_RIBOSOME gene set
genes <- get_overlap(input = multi_fora_results_ex,
                     set_name = "KEGG_RIBOSOME")

# Search genes representing KEGG_RIBOSOME overlap in all overlaps
search_results <- search_genes_in_overlaps(input = multi_fora_results_ex,
                                           genes = genes,
                                           group_name = "ribosome")

}
