% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{set_col_types}
\alias{set_col_types}
\title{Set the column types based on the column names}
\usage{
set_col_types(data_file, readr_func, ...)
}
\arguments{
\item{data_file}{Name of file to open}
}
\value{
data.frame
}
\description{
\code{set_col_types} loads the first line of
the data and sets the column type based on the column names.
Chr and Strand are factor and Start and End are integer.
normalised count columns are double (float) and
count columns are integer.
All other columns are set to the type guessed by read_tsv
}
\examples{
data <- standardise_colnames('test_data.tsv')

}
