% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_level_color.R
\name{set_level_color}
\alias{set_level_color}
\title{Set factor levels color}
\usage{
set_level_color(input, variable, levels, colors)
}
\arguments{
\item{input}{a data.frame or tibble.}

\item{variable}{character vector of length 1 representing the name of the variable to set as factor.}

\item{levels}{character vector with factor levels in desired order.}

\item{colors}{character vector with factor level colors in desired order.}
}
\value{
a tibble with variable set as factor and augmented with a variable_color column containing level colors.
}
\description{
Set factor levels color
}
\examples{
# Set donor_id as factor and assign color levels
sample_metadata <- set_level_color(input = sc_sample_metadata_ex,
                                   variable = "donor_id",
                                   levels = c("donor3", "donor2", "donor1"),
                                   colors = c("#0077b6", "#a4133c", "#29a655"))

}
