% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{split_df_ex}
\alias{split_df_ex}
\title{Split dataframe}
\format{
A data.frame with 2 columns:
\describe{
\item{spliced}{gene ID for spliced transcript (e.g.: ENSG00000005486.17 )}
\item{unspliced}{gene ID for unspliced transcript (e.g.: ENSG00000005486.17-I)}
}
}
\usage{
split_df_ex
}
\description{
Example split_df to supply to rnaseqtools::split_splici_txi.
To generate reference files for spliced and unspliced abundance estimation with alignment-free methods, see the worflow described at:
\url{https://bioconductor.org/packages/release/bioc/vignettes/eisaR/inst/doc/rna-velocity.html}.
}
\keyword{datasets}
