% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_splici_txi.R
\name{split_splici_txi}
\alias{split_splici_txi}
\title{Split txi by feature type}
\usage{
split_splici_txi(txi, split_df)
}
\arguments{
\item{txi}{txi object produced by tximport::tximport.}

\item{split_df}{data.frame or tibble with 2 columns representing feature IDs. Each column represents a separate feature type, and the features in a given row are considered representatives of the same feature.}
}
\value{
a list with distinct abundance, counts and length vectors for the 2 feature types.
}
\description{
Split txi by feature type
}
\examples{

# Split txi by splicing status (spliced / unspliced)
txi_split <- split_splici_txi(txi = txi_ex,
                              split_df = split_df_ex)

}
