% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{standardise_coltypes}
\alias{standardise_coltypes}
\title{Standardise column classes}
\usage{
standardise_coltypes(data)
}
\arguments{
\item{data}{data.frame, data to standardise}
}
\value{
data.frame
}
\description{
\code{standardise_coltypes} takes a dataframe
loaded from a RNA-seq all file and creates standard
column types. For example, Chr should be a factor even
though some times all the values will be integers
}
\examples{
data <- standardise_coltypes(data)

}
