% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{subset_to_samples}
\alias{subset_to_samples}
\title{Subset an RNA-seq dataset using a samples file}
\usage{
subset_to_samples(data, samples)
}
\arguments{
\item{data}{df rna-seq data to get counts from}

\item{samples}{df a samples df to subset the data to}
}
\value{
data.frame
}
\description{
\code{subset_to_samples} returns a data.frame of metadata,
associated with the genes of an RNA-seq dataset
}
\examples{
non_count_data <- subset_to_samples(data, samples)

}
