% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_gene_types.R
\name{summarise_gene_types}
\alias{summarise_gene_types}
\title{Summarise gene types}
\usage{
summarise_gene_types(input, gene_id, sample_id, gene_metadata)
}
\arguments{
\item{input}{data.frame or tibble with character column <gene_id> and double columns representing sample-specific gene counts.}

\item{gene_id}{character vector of length 1 or name representing gene ID. Must be a column name of input.}

\item{sample_id}{character vector of length 1 or name representing sample ID.}

\item{gene_metadata}{data.frame or tibble with character column <gene_id>, gene_symbol and gene_type. Typically produced by rnaseqtools::make_gene_metadata.}
}
\value{
a data.frame summarising the number and percentage of protein coding, mitochondrial and ribosomal genes detected in each sample.
}
\description{
Summarise gene types
}
\examples{
# Convert TPM matrix to tibble with gene IDs as column
tpm_df <- tibble::as_tibble(x = bulk_tpm_ex, rownames = "ensembl_gene_id")

# Compute mean TPMs per disease status
gene_type_summary <- summarise_gene_types(input = tpm_df,
                                          gene_id = ensembl_gene_id,
                                          sample_id = donor_id,
                                          gene_metadata = gene_metadata_ex)

}
