% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{summaryZlmFit_ex}
\alias{summaryZlmFit_ex}
\title{MAST summaryZlmFit}
\format{
An object of class \code{summaryZlmFit} (inherits from \code{list}) of length 1.
}
\usage{
summaryZlmFit_ex
}
\description{
Example summaryZlmFit object produced by MAST::summary(object = zlm, logFC = T, doLRT = T) using the workflow described at:
\url{https://bioconductor.org/packages/release/bioc/vignettes/MAST/inst/doc/MAITAnalysis.html}.
The object contains a data.table with 1000 rows and 7 columns.
}
\keyword{datasets}
