% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_DESeqResults.R
\name{tidy_DESeqResults}
\alias{tidy_DESeqResults}
\title{Tidy DESeq2 results}
\usage{
tidy_DESeqResults(input, gene_id = "ensembl_gene_id")
}
\arguments{
\item{input}{DESeqResults object produced by DESeq2::results.}

\item{gene_id}{character vector of length 1 representing gene ID. Used to create a column of gene IDs in output.}
}
\value{
a tibble with DESeq2 results.
}
\description{
Tidy DESeq2 results
}
\examples{
# Extract DESeq2 log fold change results from DESeqResults object
deseq2_results <- tidy_DESeqResults(input = DESeqResults_ex,
                                    gene_id = "ensembl_gene_id")

}
