% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_summaryZlmFit.R
\name{tidy_summaryZlmFit}
\alias{tidy_summaryZlmFit}
\title{Tidy MAST zlm summary}
\usage{
tidy_summaryZlmFit(input, coef, level = 0.95, gene_id = "ensembl_gene_id")
}
\arguments{
\item{input}{summaryZlmFit object produced by MAST::summary(object = zlm, logFC = T, doLRT = T) where zlm is a ZlmFit object. See ?MAST::\verb{summary,ZlmFit-method}.}

\item{coef}{character vector of length 1 representing the coefficient of interest for LRT.}

\item{level}{level of confidence coefficient, should be the same as in the MAST::summary call.}

\item{gene_id}{character vector of length 1 representing gene ID. Used to rename the primerid column in output.}
}
\value{
a tibble with MAST zlm summary results.
}
\description{
Tidy MAST zlm summary
}
\examples{
# Extract MAST log fold change results from summaryZlmFit object
mast_results <- tidy_summaryZlmFit(input = summaryZlmFit_ex,
                                   coef = "treatmenttreated",
                                   level = 0.95,
                                   gene_id = "ensembl_gene_id")

}
