% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_cnt_df.R
\name{transpose_cnt_df}
\alias{transpose_cnt_df}
\title{Transpose count data.frame}
\usage{
transpose_cnt_df(input, gene_id = "ensembl_gene_id", sample_id = "barcode")
}
\arguments{
\item{input}{data.frame or tibble with character column <gene_id> or <sample_id> and double columns representing sample-specific gene counts.}

\item{gene_id}{character vector of length 1 or name representing gene ID.}

\item{sample_id}{character vector of length 1 or name representing sample ID.}
}
\value{
a transposed tibble of gene counts.
}
\description{
Transpose count data.frame
}
\examples{
# Convert log2(TPM+1) matrix to tibble with gene IDs as column
log2_tpm1p_df <- tibble::as_tibble(x = sc_log2_tpm1p_ex,
                                   rownames = "ensembl_gene_id")

# Transpose log2(TPM+1) tibble with barcode as column
log2_tpm1p_df_t <- transpose_cnt_df(input = log2_tpm1p_df,
                                    gene_id = "ensembl_gene_id",
                                    sample_id = "barcode")

}
