% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools_data.R
\docType{data}
\name{txi_ex}
\alias{txi_ex}
\title{Splici txi}
\format{
An object of class \code{list} of length 4.
}
\usage{
txi_ex
}
\description{
Example txi object produced by tximport::tximport and to supply to rnaseqtools::split_splici_txi.
It contains gene-level estimates of counts, abundances and lengths for spliced and unspliced transcripts.
To generate reference files for spliced and unspliced abundance estimation with alignment-free methods, see the worflow described at:
\url{https://bioconductor.org/packages/release/bioc/vignettes/eisaR/inst/doc/rna-velocity.html}.
}
\keyword{datasets}
