% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_ids2indices.R
\name{batch_ids2indices}
\alias{batch_ids2indices}
\title{Convert gene identifiers to indices for gene sets - multiple collections}
\usage{
batch_ids2indices(
  input,
  collections,
  identifiers,
  gene_id = "ensembl_gene_id_version"
)
}
\arguments{
\item{input}{tibble with gene_id column and collections list-columns.}

\item{collections}{character vector of gene collection names; all elements must be column names of input.}

\item{identifiers}{character vector of gene identifiers.}

\item{gene_id}{character vector of length 1 representing gene ID; must be a column name of input.}
}
\value{
a list.
}
\description{
Convert gene identifiers to indices for gene sets - multiple collections
}
\examples{
# example

}
