% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msigdb_collections.R
\name{get_msigdb_collections}
\alias{get_msigdb_collections}
\title{Get MSigBD collections}
\usage{
get_msigdb_collections(
  species = "Homo sapiens",
  msigdb_collection_summary = list_msigdb_collections()
)
}
\arguments{
\item{species}{character vector of length 1 representing species. Must be a species_name returned by msigdbr::msigdbr_species.}

\item{msigdb_collection_summary}{data.frame or tibble with columns collection_name, gs_cat and gs_subcat as returned by list_msigdb_collections.}
}
\value{
a tibble with ensembl_gene_id column and collections list-columns. Each collection contains columns set_id and set_name.
}
\description{
Get MSigBD collections
}
\examples{
msigdb_collection_table = get_msigdb_collections()

}
