% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gencode_annotation.R
\name{import_gencode_annotation}
\alias{import_gencode_annotation}
\title{Import GENCODE annotation}
\usage{
import_gencode_annotation(file)
}
\arguments{
\item{file}{A GENCODE annotation file in GTF format.}
}
\value{
A tibble with GENCODE annotation.
}
\description{
Import GENCODE annotation
}
\examples{
### Do not run ###
# gencode_annotation <- import_gencode_annotation(file = "path_to_gencode_annotation.gtf")

}
