% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_msigdb_collections.R
\name{list_msigdb_collections}
\alias{list_msigdb_collections}
\title{List MSigDB collections}
\usage{
list_msigdb_collections(db_species = "Hs")
}
\arguments{
\item{db_species}{character vector of length 1 representing species abbreviation for database query. Must be "Hs" or "HS" for human and "Mm" or "MM" for mouse databases.}
}
\value{
a tibble summarizing available MSigDB collections.
}
\description{
List MSigDB collections
}
\examples{
msigdb_collection_summary = list_msigdb_collections()

}
