% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_msigdb_collections.R
\name{list_msigdb_collections}
\alias{list_msigdb_collections}
\title{List MSigBD collections}
\usage{
list_msigdb_collections()
}
\value{
a tibble summarizing available MSigDB collections with columns collection_name, gs_cat and gs_subcat.
}
\description{
List MSigBD collections
}
\examples{
msigdb_collection_summary = list_msigdb_collections()

}
