% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fgsea_pathways.R
\name{make_fgsea_pathways}
\alias{make_fgsea_pathways}
\title{Make fgsea pathways}
\usage{
make_fgsea_pathways(input, collections, gene_id = "ensembl_gene_id_version")
}
\arguments{
\item{input}{tibble with gene_id column and collections list-columns.}

\item{collections}{character vector of gene collection names. All elements must be column names of input.}

\item{gene_id}{character vector of length 1 representing gene ID. Must be a column name of input.}
}
\value{
a list of collection-specific sublists containing gene set-specific character vectors of gene IDs.
}
\description{
Make fgsea pathways
}
\examples{
### Do not run ###
# collections <- make_fgsea_pathways(input = msigdb_collection_table,
#                                    collections = collection_names,
#                                    gene_id = "ensembl_gene_id")

}
