% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_fgsea.R
\name{multi_fgsea}
\alias{multi_fgsea}
\title{Run fgsea::fgsea on multiple collections}
\usage{
multi_fgsea(
  input,
  collections,
  stats,
  gene_id = "ensembl_gene_id_version",
  min_set_size = 1,
  max_set_size = length(stats) - 1,
  padj_threshold = Inf,
  ...
)
}
\arguments{
\item{input}{tibble with gene_id column and collections list-columns.}

\item{collections}{character vector of gene collection names. All elements must be column names of input.}

\item{stats}{named vector of gene-level stats. Names should be the same as in gene_id.}

\item{gene_id}{character vector of length 1 representing gene ID. Must be a column name of input.}

\item{min_set_size}{minimal size of a gene set to test. All gene sets below the threshold are excluded.}

\item{max_set_size}{maximal size of a gene set to test. All gene sets above the threshold are excluded.}

\item{padj_threshold}{padj threshold. All gene sets with padj equal or above the threshold are filtered out.}

\item{...}{optional arguments passed to fgsea::fgsea.}
}
\value{
a tibble with results of fgsea::fgsea runs on multiple collections.
}
\description{
Run fgsea::fgsea on multiple collections
}
\examples{
### Do not run ###
# fgsea_results <- multi_fgsea(input = msigdb_collection_table,
#                              collections = collection_names,
#                              stats = stats,
#                              gene_id = "ensembl_gene_id")

}
