% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_fora.R
\name{multi_fora}
\alias{multi_fora}
\title{Run fgsea::fora on multiple collections}
\usage{
multi_fora(
  input,
  collections,
  genes,
  universe,
  gene_id = "ensembl_gene_id_version",
  min_set_size = 1,
  max_set_size = NULL,
  padj_threshold = Inf,
  ...
)
}
\arguments{
\item{input}{tibble with gene_id column and collections list-columns.}

\item{collections}{character vector of gene collection names. All elements must be column names of input.}

\item{genes}{set of query genes.}

\item{universe}{a universe from which genes were selected.}

\item{gene_id}{character vector of length 1 representing gene ID. Must be a column name of input.}

\item{min_set_size}{minimal size of a gene set to test. All gene sets below the threshold are excluded.}

\item{max_set_size}{maximal size of a gene set to test. All gene sets above the threshold are excluded.}

\item{padj_threshold}{padj threshold. All gene sets with padj equal or above the threshold are filtered out.}

\item{...}{optional arguments passed to fgsea::fora.}
}
\value{
a tibble with results of fgsea::fora runs on multiple collections.
}
\description{
Run fgsea::fora on multiple collections
}
\examples{
### Do not run ###
# fora_results <- multi_ora(input = msigdb_collection_table,
#                           collections = collection_names,
#                           genes = genes,
#                           universe = universe,
#                           gene_id = "ensembl_gene_id")

}
