% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools-package.R
\docType{package}
\name{rnaseqtools-package}
\alias{rnaseqtools}
\alias{rnaseqtools-package}
\title{rnaseqtools: Downstream RNA-Seq Data Analysis and Visualization}
\description{
A collection of tools for RNA-seq data analysis, including gene count manipulation and summarization, gene set testing and data visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benvallin/rnaseqtools}
  \item Report bugs at \url{https://github.com/benvallin/rnaseqtools/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Vallin \email{ben.vallin@gmail.com} (\href{https://orcid.org/0009-0008-4907-4330}{ORCID})

}
\keyword{internal}
