% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqtools-package.R
\docType{package}
\name{rnaseqtools-package}
\alias{rnaseqtools}
\alias{rnaseqtools-package}
\title{rnaseqtools: RNA-seq analysis tools}
\description{
Helper functions to facilitate RNA-seq analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benvallin/rnaseqtools}
  \item Report bugs at \url{https://github.com/benvallin/rnaseqtools/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Vallin \email{ben.vallin@gmail.com} (\href{https://orcid.org/0009-0008-4907-4330}{ORCID})

}
\keyword{internal}
