\name{RNASeqTools-package}
\alias{RNASeqTools-package}
\alias{RNASeqTools}
\docType{package}
\title{RNASeqTools is a Collection of Helpful Functions for RNA-Seq analysis}
\description{
  \code{RNASeqTools} contains functions for manipulation expression
  data, both as dataframes and as objects of class \code{CountDataSet}
  from the \code{DESeq} package. 
}
\details{
\tabular{ll}{
Package: \tab RNASeqTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-01-19\cr
License: \tab GPL 1.0\cr
Depends: \tab methods\cr
}

}
\author{
Vince Buffalo <vsbuffalo@ucdavis.edu>

Maintainer: Vince Buffalo <vsbuffalo@ucdavis.edu>
}
\references{
}
\keyword{ package }
\seealso{
}

